Move URL redundants methods to an extension
Because of this, to have following methods available in `Conn`, `:url`
extensions needs first to be loaded:

- base_url
- path
- full_path
- url
- router_params
- params
